# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Your API Description",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "License Name",
            "url": "http://www.example.com/license"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/dashboard/subscription/client/{id}": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Retrieve paginated list of subscriptions for a specific client",
                "operationId": "4d9f271c14f031ccfeb740a67374ed7f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Client's ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client subscriptions retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "client subscriptions fetched successfully"
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Subscription"
                                            }
                                        },
                                        "user_details": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john.doe@example.com"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "1234567890"
                                                },
                                                "age": {
                                                    "type": "integer",
                                                    "example": 30
                                                },
                                                "weight": {
                                                    "type": "integer",
                                                    "example": 70
                                                },
                                                "height": {
                                                    "type": "integer",
                                                    "example": 180
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No subscriptions found for the given client_id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No subscriptions found for the given client_id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/dashboard/subscription": {
            "post": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Create a new subscription",
                "operationId": "6fd0892566864b63c9e2fb80866e6030",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nutrition_coach_id": {
                                        "description": "Must exist in nutrition coaches table",
                                        "type": "integer"
                                    },
                                    "workout_coach_id": {
                                        "description": "Must exist in workout coaches table",
                                        "type": "integer"
                                    },
                                    "client_id": {
                                        "description": "Must exist in clients table",
                                        "type": "integer"
                                    },
                                    "sale_id": {
                                        "description": "Must exist in sales table",
                                        "type": "integer"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-01-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-12-31"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subscription created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Subscription"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation errors"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/dashboard/subscription/{id}": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Retrieve a specific subscription",
                "operationId": "0de8e00940b4bd73dee675b09cfe9d30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Subscription's ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subscription retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Subscription"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Subscription not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Subscription not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Update a specific subscription",
                "operationId": "d16aadbc9a3bb9b705b6aaf0be5bc048",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Subscription's ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nutrition_coach_id": {
                                        "description": "Must exist in nutrition coaches table",
                                        "type": "integer"
                                    },
                                    "workout_coach_id": {
                                        "description": "Must exist in workout coaches table",
                                        "type": "integer"
                                    },
                                    "client_id": {
                                        "description": "Must exist in clients table",
                                        "type": "integer"
                                    },
                                    "sale_id": {
                                        "description": "Must exist in sales table",
                                        "type": "integer"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-01-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-12-31"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subscription updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Subscription"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation errors"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Subscription not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Subscription not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Subscription": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nutrition_coach_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "workout_coach_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "client_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "sale_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "duration": {
                        "type": "integer",
                        "example": 30
                    },
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "package": {
                        "type": "string",
                        "example": "Basic"
                    },
                    "started_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2022-01-01"
                    },
                    "paid_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "freeze_start_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2022-01-01"
                    },
                    "freeze_duration": {
                        "type": "integer",
                        "example": 7
                    },
                    "paid_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2022-01-01"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2022-01-01"
                    }
                },
                "type": "object"
            },
...
        }
    }
}